backend/requirements.txt
fastapi
mangum
openai
jinja2
pydantic
python-dotenv
python-multipart
uvicorn
